requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
g4f[all]==0.3.1.9
deep-translator==1.11.4
pymongo==4.7.1
pytz==2024.1
Flask==3.0.3
safoneapi==1.0.63
pytube==15.0.0
psutil==5.9.8
telegraph==2.2.0
requests
TgCrypto